SINGA-349 Create layer operations for autograd
1. Modified the new designed API, fix bugs.
   It can work in my computer.

2. Do some codes relocation:
   - Integrate operations into autograd.py
   - Delete extra files
   - Delete operations in tensor.py
