+++ python/singa/tensor.py
@@ -97,7 +97,7 @@ class Tensor(object):
             copy_from_numpy(self.data, data)
         elif isinstance(data, CTensor):
             self.data = data
-            assert data.device == device, 'not the same device'
+            assert data.device() == device, 'not the same device'
         else:
             self.data = CTensor(list(shape), device, dtype)
 
@@ -1356,225 +1356,3 @@ class Dummy(Operation):
         self.src = []
         self.y_id2idx = {id(tensor): 0}
         self.requires_grad = False
-
-
-class ReLU(Operation):
-
-    def forward(self, x):
-        '''
-        Args:
-            x(CTensor): input tensor
-
-        Returns:
-            a new CTensor whose element y = x if x >= 0; otherwise 0;
-        '''
-        self.input = x
-        return singa.ReLU(x)
-
-    def backward(self, dy):
-        '''
-        Args:
-            dy(CTensor): dL / dy
-
-        Returns:
-            dx(CTensor): dL / dx = dy if x >= 0; otherwise 0;
-        '''
-        dx = singa.GTFloat(self.input, 0.0)
-        return singa.__mul__(dy, dx)
-
-
-def relu(x):
-    return ReLU()(x)[0]
-
-
-class Matmul(Operation):
-    '''For matrix multiplication'''
-
-    def forward(self, x, w):
-        '''Do forward propgation.
-
-        Store the x(or w) if w(or x) requires gradient.
-
-        Args:
-            x (CTensor): matrix
-            w (CTensor): matrix
-
-        Returns:
-            a CTensor for the result
-        '''
-        self.input = (x, w)
-        return singa.Mult(x, w)
-
-    def backward(self, dy):
-        '''
-        Args:
-            dy (CTensor): data for the dL / dy, L is the loss
-
-        Returns:
-            a tuple for (dx, dw)
-        '''
-        return singa.Mult(dy, self.input[1].T()), \
-            singa.Mult(self.input[0].T(), dy)
-
-
-def matmul(x, w):
-    return Matmul()(x, w)[0]
-
-
-class AddBias(Operation):
-    '''
-    Add Bias to each row / column of the Tensor, depending on the parameter axis.
-    '''
-
-    def __init__(self, axis=0):
-        '''
-        To indicate the calculation axis, 0 for row, 1 for column.
-
-        Args:
-            axis: 0 or 1, default is 0.
-        '''
-        self.axis = axis
-
-    def forward(self, x, b):
-        '''
-        Args:
-            x: matrix.
-            b: bias to be added.
-
-        Return:
-            the result Tensor
-        '''
-        if self.axis == 0:
-            singa.AddRow(b, x)
-        elif self.axis == 1:
-            singa.AddColumn(b, x)
-        return x
-
-    def backward(self, dy):
-        '''
-        Args:
-            dy (CTensor): data for the dL / dy, L is the loss.
-
-        Return:
-            a tuple for (db, dx), db is data for dL / db, dx is data
-            for dL / dx.
-        '''
-        if self.axis == 0:
-            return dy, singa.Sum(dy, 0)
-        elif self.axis == 1:
-            return dy, singa.Sum(dy, 0)
-
-
-def add_bias(x, b, axis=0):
-    return AddBias(axis)(x, b)[0]
-
-
-class SoftMax(Operation):
-    '''
-    Apply SoftMax for each row of the Tensor or each column of the Tensor
-    according to the parameter axis.
-    '''
-
-    def __init__(self, axis=0):
-        self.axis = axis
-
-    def forward(self, x):
-        '''
-        Args:
-            x(data): the input 1d or 2d tensor
-
-        Returns:
-            the result Tensor
-        '''
-        if self.axis == 1:
-            x = x.T()
-        self.output = singa.SoftMax(x)
-        if self.axis == 0:
-            return self.output
-        elif self.axis == 1:
-            return self.output.T()
-
-    def backward(self, dy):
-        '''
-        Args:
-            dy (CTensor): data for the dL / dy, L is the loss
-
-        Returns:
-            dx (Ctensor): data for the dL / dx, L is the loss,
-            x is the input of current Opertion
-        '''
-        # calculations are made on numpy array
-        if self.axis == 1:
-            dy = dy.T()
-        grad = ctensor2numpy(dy)
-        output = ctensor2numpy(self.output)
-        out_1 = np.einsum('ki,ki->ki', grad, output)
-        medium_out = np.einsum('ki,kj->kij', output, output)
-        out_2 = np.einsum('kij,kj->ki', medium_out, grad)
-        out = out_1 - out_2
-        dx = CTensor(out_1.shape)
-        dx.CopyFloatDataFromHostPtr(out.flatten())
-        if self.axis == 0:
-            return dx
-        elif self.axis == 1:
-            return dx.T()
-
-
-def soft_max(x, axis=0):
-    return SoftMax(axis)(x)[0]
-
-
-class CrossEntropy(Operation):
-    '''
-    Calculte CrossEntropy loss for a batch of training data.
-
-    '''
-
-    def forward(self, x, t):
-        '''
-        Args:
-            x (CTensor): 1d or 2d tensor, the prediction data(output) of current network.
-            t (CTensor): 1d or 2d tensor, the target data for training.
-
-        Returns:
-            loss (CTensor): scalar.
-        '''
-        loss = CTensor((1,))
-        loss_data = -singa.SumAsFloat(singa.__mul__(t, singa.Log(x)))
-        loss.SetFloatValue(loss_data / x.shape()[0])
-        self.x = x
-        self.t = t
-        self.input = (x, t)
-        return loss
-
-    def backward(self, dy=1.0):
-        '''
-        Args:
-            dy (float or CTensor): scalar, accumulate gradient from outside of current network, usually
-            equal to 1.0
-
-        Returns:
-            dx (CTensor): data for the dL /dx, L is the loss, x is the output of current network.
-            note that this is true for dy = 1.0
-        '''
-        dx = singa.__div__(self.t, self.x)
-        dx *= float(-1 / self.x.shape()[0])
-        if isinstance(dy, float):
-            # dtype of dy: float
-            dx *= dy
-            return dx, None
-        elif isinstance(dy, CTensor):
-            pass  # TODO, broadcast elementwise multiply seems not support
-
-
-def cross_entropy(y, t):
-    return CrossEntropy()(y, t)[0]
-
-
-def ctensor2numpy(x):
-    '''
-    To be used in SoftMax Operation.
-    Convert a singa_tensor to numpy_tensor.
-    '''
-    np_array = x.GetFloatValue(int(x.Size()))
-    return np_array.reshape(x.shape())
